% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFFiles.R
\name{fixColnamesForBigQuery}
\alias{fixColnamesForBigQuery}
\title{Fix the column names from a data.frame for insertion to BigQuery}
\usage{
fixColnamesForBigQuery(df)
}
\arguments{
\item{df}{a data.frame or data.frame-like object}
}
\description{
BigQuery requires that names be letters, underscores, or numbers.
This function returns a data.frame with column names that conform
to that standard.
}

