% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFFiles.R
\name{getNBLExomeMutationMAF}
\alias{getNBLExomeMutationMAF}
\title{Get the Neuroblastoms Exome-based Somatic MAF file}
\usage{
getNBLExomeMutationMAF()
}
\description{
Get the Neuroblastoms Exome-based Somatic MAF file
}

