% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFFiles.R
\name{getWTExomeMutationMAF}
\alias{getWTExomeMutationMAF}
\title{Get the Wilms Tumor Exome-based Somatic MAF file}
\usage{
getWTExomeMutationMAF()
}
\description{
Get the Wilms Tumor Exome-based Somatic MAF file
}

